# perf: make silverback --help faster