--- a/setup.cfg
+++ b/setup.cfg
@@ -1,5 +1,6 @@
 [flake8]
 max-line-length = 100
+ignore = E704,W503,PYD002
 exclude =
 	venv*
 	.eggs