--- a/silverback/__init__.py
+++ b/silverback/__init__.py
@@ -1,6 +1,24 @@
-from .exceptions import CircuitBreaker, SilverbackException
-from .main import SilverbackBot
-from .state import StateSnapshot
+def __getattr__(name: str):
+    if name == "CircuitBreaker":
+        from .exceptions import CircuitBreaker
+
+        return CircuitBreaker
+
+    elif name == "SilverbackException":
+        from .exceptions import SilverbackException
+
+        return SilverbackException
+
+    elif name == "SilverbackBot":
+        from .main import SilverbackBot
+
+        return SilverbackBot
+
+    elif name == "StateSnapshot":
+        from .state import StateSnapshot
+
+        return StateSnapshot
+
 
 __all__ = [
     "StateSnapshot",