### What I did

Cuts the time down. It is twice as fast for me now to run `silverback --help`.

### How I did it

* Any CLI module has to be more careful on its top-level imports
* Root init file of package needs to also be careful, using a `__getattr__` there.

### How to verify it

```sh
time silverback --help
```


Before:

```
silverback  4.13s user 0.83s system 93% cpu 5.280 total
```

After:

```
silverback  1.63s user 0.28s system 83% cpu 2.297 total
```

Results may vary.

### Checklist

- [ ] Passes all linting checks (pre-commit and CI jobs)
- [ ] New test cases have been added and are passing
- [ ] Documentation has been updated
- [ ] PR title follows [Conventional Commit](https://www.conventionalcommits.org/en/v1.0.0/) standard (will be automatically included in the changelog)