@dtdang 2024-10-30
Looks like some of the CLI commands that depend on ApePay are causing errors in the new PR since we refactored `BaseInterfaceModel` to utils so we'll need to fix this first before merging the PR.

`ImportError: cannot import name 'BaseInterfaceModel' from 'ape.types'`