@antazoey 2024-10-30
> `ImportError: cannot import name 'BaseInterfaceModel' from 'ape.types'`

`BaseInterfaceModel` was never exported in `ape.types`; this was a mistake