@dtdang 2024-10-30
> > `ImportError: cannot import name 'BaseInterfaceModel' from 'ape.types'`
> 
> `BaseInterfaceModel` was never exported in `ape.types`; this was a mistake it was only coincidentally imported there

Looks like I just needed to update the repo on my end, my bad. Testing the rest of the commands now but everything is looking good since pulling the latest repos.