@antazoey 2024-10-30
Well I am moving all the base model stuff to types since that is where yall went lookin for it: https://github.com/ApeWorX/ape/pull/2353
And so in 0.9 youll have to put that back or wait for the release of Ape where itll work in both