@dtdang 2024-11-04
CLI is still working as intended