@antazoey 2024-11-04
> The type annotation `PlatformClient` was something I was using internally, we don't really have the CLI tested here so I want to make sure nothing was breaking

I am quite tempted to add more tests to silverback.